repo resubido

+++ Division/div/div.go
@@ -1,7 +1,7 @@
 package div //si algo empieza con minuscula es privado al package
 import "errors"
 
-func Division(a int, b int) (int, error) {
+func Division(a float64, b float64) (float64, error) {
 	if b == 0 {
 		return 0, errors.New("No se puede dividir")
 	}
